Behavior of the split view of the graphic window is fixed
git-svn-id: svn+ssh://svn.osdn.jp/svnroot/alchemusica/trunk@169 19e2383c-df9d-4eb5-8775-ae43218f5bcb

+++ Classes/GraphicClientView.h
@@ -37,8 +37,8 @@ enum {
     BOOL isDragging;
     BOOL isLoupeDragging;
 	BOOL autoScaleOnResizing;
-    unsigned int initialModifierFlags;
-    unsigned int currentModifierFlags;
+    NSUInteger initialModifierFlags;
+    NSUInteger currentModifierFlags;
     NSMutableArray *selectPoints;
     NSBezierPath *selectionPath;
 	NSRect initialSelectionRect;
